Session6_AgenticArchitecture/requirements.txt
Flask>=2.2.0
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generative-ai>=0.1.0
mcp>=0.1.0
